+++ src/store/sagas/musicSaga.tsx
@@ -19,21 +19,21 @@ import { setItem } from '../../utils/localstorage';
 import history from '../../utils/history';
 import { IMusic } from '../../@types/music';
 
-function* getMusicList({ payload: keyword }: any) {
-  const beforeKeyword: string = yield select((state) => state.music.keyword);
+// function* getMusicList({ payload: keyword }: any) {
+//   const beforeKeyword: string = yield select((state) => state.music.keyword);
 
-  try {
-    if (keyword === beforeKeyword) throw new Error('동일한 키워드 검색');
+//   try {
+//     if (keyword === beforeKeyword) throw new Error('동일한 키워드 검색');
 
-    yield put(setKeyword(keyword));
-    const data: IMusic[] = yield call(fetchPlayList, `${keyword} 플레이리스트`);
+//     yield put(setKeyword(keyword));
+//     const data: IMusic[] = yield call(fetchPlayList, `${keyword} 플레이리스트`);
 
-    yield put(getMusicListSuccess(data));
-    history.replace(`/search?query=${keyword}`);
-  } catch (e) {
-    // yield put(getMusicListFail(e));
-  }
-}
+//     yield put(getMusicListSuccess(data));
+//     history.replace(`/search?query=${keyword}`);
+//   } catch (e) {
+//     // yield put(getMusicListFail(e));
+//   }
+// }
 
 function* addDjplaylist({ payload: selectedMusic }: any) {
   const djPlaylist: IMusic[] = yield select((state) => state.music.djPlaylist);
@@ -105,7 +105,7 @@ function* swapMusicList({ payload: route }: any) {
 }
 
 export default function* watchMusic() {
-  yield takeEvery(GET_MUSICLIST, getMusicList);
+  // yield takeEvery(GET_MUSICLIST, getMusicList);
   yield takeEvery(ADD_DJPLAYLIST, addDjplaylist);
   yield takeEvery(DELETE_DJPLAYLIST, deleteDjplaylist);
   yield takeLatest(PREV_MUSIC, prevMusic);
